requirements.txt
pygame
Pillow
# Optional: tzlocal
# Optional: watchdog
# Optional (video playback): moviepy
